html에 그래프 데이터 적용하여 완성& 내가 만든 함수를 API테이블로 이동
