+++ corona_map/urls.py
@@ -31,7 +31,7 @@ urlpatterns = [
     path('data-init/', views.call_data_init, name='data-init'),
 
     # data test : 수녕
-    path('cois-test/', views.infection_city_all_values, name='cois-test'),
+    # path('cois-test/', views.infection_city_all_values, name='cois-test'),
 
 
     # loacalhost:8080/board-list : 코로나 뉴스게시판 리스트 페이지
