+++ corona_map/views.py
@@ -9,14 +9,11 @@ import corona_map.MongoDbManager as comong
 from corona_map.Api import Infection_city, Infection_status, Infection_by_age_gender, News_board
 import pymongo
 import datetime
-
 from corona_map.Api.Gugun_status import get_seoul_data_list, init_gugun_data
-
-
-import corona_map.MongoDbManager as comong
 from corona_map.Api.Infection_city import infection_city
 from corona_map.Api.Infection_by_age_gender import infection_by_age_gender
 from corona_map.Api.Infection_status import infection_status
+from corona_map.Api.main_data_graph_function import infection_city_all_values, infection_city_oneday_values, infection_all_value, infection_oneday_value, infection_by_age_all_value, infection_by_gender_all_value
 
 from corona_map.Api.data_init import seoul_data_init
 from corona_map.Api.Gugun_status_calc import get_seoul_calc_data_list
@@ -65,6 +62,7 @@ def call_gugun_info(request):
 #     return infection_city_all_values_list
 
 
+
 # infection_state collection 전국 코로나 현황 수 get함수
 def infection_state_all_value():
     now = datetime.datetime.now()
@@ -87,172 +85,6 @@ def infection_state_all_value():
     return item_dict
 
 
-'''
-# 작성자 : 최수녕
-# 함수 설명 : infection_city 테이블에서 id의 최댓값을 구하는 함수(가장 최근 날짜 구하기)
-# 리턴값 : id의 최댓값(숫자)
-# 분류 : 추후 API폴더로 이동 예정
-'''
-def infection_city_max_id():
-    max_id = comong.Infection_City().get_aggregate_users_from_collection([{'$group':{'_id':'null','total':{'$max':'$id'}}},{'$project':{'_id':0,'total':1}}])
-    max_id = list(max_id)[0]['total']
-    return max_id
-
-
-'''
-# 작성자 : 최수녕
-# 함수 설명 : 시도별 총확진자 구하는 함수
-# 리턴값 : 시/도 이름 list , 시/도에 따른 총확진자 list
-# 분류 : 추후 API폴더로 이동 예정
-'''
-def infection_city_all_values():
-    max_date = infection_city_max_id()
-    infection_date_data = comong.Infection_City().get_particular_users_from_collection({'$and':[{'id':max_date},{'gubun':{'$nin':['합계','검역']}}]},{'gubun': 1, 'defcnt': 1, '_id': 0})
-    i_city_all_key = []
-    i_city_all_value = []
-    for infection_data in infection_date_data:
-        i_city_all_key.append(infection_data['gubun'])
-        i_city_all_value.append(infection_data['defcnt'])
-    context = {'i_city_all_key': i_city_all_key, 'i_city_all_value': i_city_all_value}
-    return context
-
-
-'''
-# 작성자 : 최수녕
-# 함수 설명 : 시도별 일별 확진자 구하는 함수
-# 리턴값 : 시/도 이름 list , 시/도에 따른 확진자(전일대비 증감 수) list
-# 분류 : 추후 API폴더로 이동 예정
-# 현황 : 구현중
-'''
-def infection_city_oneday_values():
-    max_date = infection_city_max_id()
-    infection_date_data = comong.Infection_City().get_particular_users_from_collection({'$and':[{'id':max_date},{'gubun':{'$nin':['합계','검역']}}]},{'gubun': 1, 'incdec': 1, '_id': 0})
-    i_city_oneday_key = []
-    i_city_oneday_value = []
-    for infection_data in infection_date_data:
-        i_city_oneday_key.append(infection_data['gubun'])
-        i_city_oneday_value.append(infection_data['incdec'])
-    context = {'i_city_oneday_key': i_city_oneday_key, 'i_city_oneday_value': i_city_oneday_value}
-    return context
-
-
-'''
-# 작성자 : 최수녕
-# 함수 설명 : 전국 일별 코로나 총확진자 데이터 구하는 함수
-# 리턴값 : 날자값 list, 날자에 따른 일별 확진자 list
-# 분류 : 추후 API폴더로 이동 예정
-# 현황 : 구현중
-'''
-def infection_all_value():
-    now = datetime.datetime.now()
-    nowDate = int(now.strftime('%Y%m%d'))
-    pastday_at_this_time = datetime.datetime.now() - datetime.timedelta(days=7)
-    pastDate = int(pastday_at_this_time.strftime('%Y%m%d'))
-
-    infection_date_data = comong.Infection_Status().get_particular_users_from_collection({'$and':[{'id':{'$gte':pastDate}},{'id':{'$lte':nowDate}}]},{'id': 1, 'decidecnt': 1, '_id': 0})
-    i_state_all_key = []
-    i_state_all_value = []
-    for infection_data in infection_date_data:
-        i_state_all_key.append(infection_data['id'])
-        i_state_all_value.append(infection_data['decidecnt'])
-    context = {'i_state_all_key': i_state_all_key, 'i_state_all_value': i_state_all_value}
-    return context
-
-
-'''
-# 작성자 : 최수녕
-# 함수 설명 : 전국 일별(일별 순수 확진자) 코로나 확진자 데이터 구하는 함수
-# 리턴값 : 날자값 list, 날자에 따른 일별 확진자 list
-# 분류 : 추후 API폴더로 이동 예정
-'''
-def infection_oneday_value():
-    now = datetime.datetime.now()
-    nowDate = int(now.strftime('%Y%m%d'))
-    pastday_at_this_time = datetime.datetime.now() - datetime.timedelta(days=7)
-    pastDate = int(pastday_at_this_time.strftime('%Y%m%d'))
-
-    infection_date_data = comong.Infection_Status().get_particular_users_from_collection(
-        {'$and': [{'id': {'$gte': pastDate}}, {'id': {'$lte': nowDate}}]},
-        {'decidecnt': 1, 'id': 1, '_id': 0})
-
-    infection_data_list = list(infection_date_data)
-    oneday_value_list = []
-    oneday_key_list = []
-    for i in range(0, len(infection_data_list) - 1):
-        decidecnt_oneday_data = int(infection_data_list[i]['decidecnt'] - infection_data_list[i + 1]['decidecnt'])
-        id_oneday_data = infection_data_list[i]['id']
-        oneday_value_list.append(decidecnt_oneday_data)
-        oneday_key_list.append(id_oneday_data)
-
-    context = {'oneday_value_list': oneday_value_list, 'oneday_key_list': oneday_key_list}
-    return context
-
-
-'''
-# 작성자 : 최수녕
-# 함수 설명 : 나이별 치명률 평균값 구하는 함수
-# 리턴값 : 성별(여성/남성) list , 성별에 따른 치명률 list
-# 분류 : 추후 API폴더로 이동 예정
-'''
-def infection_by_age_all_value():
-    infection_date_data = comong.Infection_By_Age_Gender().get_aggregate_users_from_collection(
-        [
-            {'$match': {'$and': [{'gubun': {'$not': {'$regex': '여성'}}}, {'gubun': {'$not': {'$regex': '남성'}}}]}},
-            {'$group': {'_id': '$gubun', 'mean_criticalrate': {'$avg': '$criticalrate'}}},
-            {'$sort': {'mean_criticalrate': -1}},
-            {
-                '$project': {
-                    'gubun': '$_id',
-                    'mean_criticalrate': 1,
-                    '_id': 0
-                }
-            }
-        ]
-    )
-    infection_by_age_all_value_list = list(infection_date_data)
-    age_key_list = []
-    age_value_list = []
-    for infection_by_age_all_value_dict in infection_by_age_all_value_list:
-        age_key_list.append(infection_by_age_all_value_dict['gubun'])
-        age_value_list.append(infection_by_age_all_value_dict['mean_criticalrate'])
-
-    context = {'age_key_list': age_key_list, 'age_value_list': age_value_list}
-
-    return context
-
-
-'''
-# 작성자 : 최수녕
-# 함수 설명 : 성별 치명률 평균값 구하는 함수
-# 리턴값 : 성별(여성/남성) list , 성별에 따른 치명률 list
-# 분류 : 추후 API폴더로 이동 예정
-'''
-def infection_by_gender_all_value():
-    infection_date_data = comong.Infection_By_Age_Gender().get_aggregate_users_from_collection(
-        [
-            {'$match': {'$or': [{'gubun': {'$regex': '여성'}}, {'gubun': {'$regex': '남성'}}]}},
-            {'$group': {'_id': '$gubun', 'mean_criticalrate': {'$avg': '$criticalrate'}}},
-            {'$sort': {'mean_criticalrate': -1}},
-            {
-                '$project': {
-                    'gubun': '$_id',
-                    'mean_criticalrate': 1,
-                    '_id': 0
-                }
-            }
-        ]
-    )
-    infection_by_gender_all_value_list = list(infection_date_data)
-    age_key_list = []
-    age_value_list = []
-    for infection_by_age_all_value_dict in infection_by_gender_all_value_list:
-        age_key_list.append(infection_by_age_all_value_dict['gubun'])
-        age_value_list.append(infection_by_age_all_value_dict['mean_criticalrate'])
-
-    context = {'gender_key_list': age_key_list, 'gender_value_list': age_value_list}
-    return context
-
-
 
 # 템플릿 적용
 def cois_main(request):
@@ -261,48 +93,46 @@ def cois_main(request):
 
     # 지역별 코로나 총확진자 현황
     item_list_result = infection_city_all_values()
-    gubunNames = item_list_result['i_city_all_key']
-    barPlotVals = item_list_result['i_city_all_value']
+    barCityAllKeys = item_list_result['i_city_all_key']
+    barCityAllVals = item_list_result['i_city_all_value']
 
     # 지역별 코로나 일별확진자 현황
     item_city_oneday_result = infection_city_oneday_values()
-    i_city_oneday_key = item_city_oneday_result['i_city_oneday_key']
-    i_city_oneday_value = item_city_oneday_result['i_city_oneday_value']
+    barCityOnedayKeys = item_city_oneday_result['i_city_oneday_key']
+    barCityOnedayVals = item_city_oneday_result['i_city_oneday_value']
 
 
     # 날자별 코로나 총확진자 현황
     item_state_all_result = infection_all_value()
-    lineChartVals = item_state_all_result['i_state_all_key']
-    dateTimes = item_state_all_result['i_state_all_value']
+    lineAllKeys = item_state_all_result['i_state_all_key']
+    lineAllVals = item_state_all_result['i_state_all_value']
 
-    # 날자별 코로나 일변확진자 현황
+    # 날자별 코로나 일별확진자 현황
     item_i_oneday_result = infection_oneday_value()
-    oneday_key_list = item_i_oneday_result['oneday_key_list']
-    oneday_value_list = item_i_oneday_result['oneday_value_list']
+    lineOnedayKeys = item_i_oneday_result['oneday_key_list']
+    lineOnedayVals = item_i_oneday_result['oneday_value_list']
 
     ##################################################################################
 
     # 성별, 연령별
-    # 연령별 데이터 가져오기
+    # 연령별 치명률 현황
     age_result_dict = infection_by_age_all_value()
-    # 성별 데이터 가져오기
-    gender_result_dict = infection_by_gender_all_value()
-
-   # 연령별 치명률 시각화 데이터
-    oldPlotVals = age_result_dict['age_value_list']
     oldGubunNames = age_result_dict['age_key_list']
+    oldPlotVals = age_result_dict['age_value_list']
 
 
-    # 성별 치명률 시각화 데이터
-    genderPlotVals = gender_result_dict['gender_value_list']
+    # 성별 치명률 현황
+    gender_result_dict = infection_by_gender_all_value()
     genderGubunNames = gender_result_dict['gender_key_list']
+    genderPlotVals = gender_result_dict['gender_value_list']
 
-    context = {'barPlotVals': barPlotVals, 'gubunNames': gubunNames,
-               'lineChartVals': lineChartVals, 'dateTimes': dateTimes, 'oldPlotVals': oldPlotVals, 'oldGubunNames': oldGubunNames,
-               'genderPlotVals': genderPlotVals, 'genderGubunNames': genderGubunNames, 'decideCnt': in_st_dict['decidecnt'],'clearCnt': in_st_dict['clearcnt'],'examCnt': in_st_dict['examcnt'],'deathCnt': in_st_dict['deathcnt']}
+    context = {'barCityAllVals': barCityAllVals, 'barCityAllKeys': barCityAllKeys,
+               'lineAllVals': lineAllVals, 'lineAllKeys': lineAllKeys, 'oldPlotVals': oldPlotVals, 'oldGubunNames': oldGubunNames,
+               'genderPlotVals': genderPlotVals, 'genderGubunNames': genderGubunNames, 'decideCnt': in_st_dict['decidecnt'],
+               'clearCnt': in_st_dict['clearcnt'],'examCnt': in_st_dict['examcnt'],'deathCnt': in_st_dict['deathcnt'],
+               'lineOnedayKeys': lineOnedayKeys, 'lineOnedayVals': lineOnedayVals,
+               'barCityOnedayKeys': barCityOnedayKeys, 'barCityOnedayVals': barCityOnedayVals}
 
-
-    print(dateTimes)
     return render(request, 'corona_map/index.html', context)
 
 
@@ -331,12 +161,6 @@ def infection_city_gubun_defcnt():
 # infection_city_gubun_defcnt() 함수 사용
 def folium_page(request):
     # mongodb collection infection_city에 api request해서 데이터 저장.
-    # print(Infection_city.infection_city())
-    # print(News_board.news_board_list())
-    # print(Infection_by_age_gender.infection_by_age_gender())
-    # print(Infection_status.infection_status())
-
-
     # print(Infection_city.infection_city())
     # print(News_board.news_board_list())
     # print(Infection_by_age_gender.infection_by_age_gender())
